requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.10
python-dotenv==1.0.0
requests==2.31.0
httpx==0.25.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.110.0
openai==1.30.0
pytest==7.4.3
pytest-flask==1.3.0
pytz==2024.1
